+++ starsky/starsky.foundation.database/Helpers/RunMigrations.cs
@@ -82,6 +82,7 @@ public static class RunMigrations
 				innerExceptionMessage.Append(innerException.Message);
 				innerExceptionMessage.Append(innerException.InnerException?.Message);
 				innerExceptionMessage.Append(innerException.InnerException?.Data);
+				innerExceptionMessage.Append(innerException.InnerException?.StackTrace);
 			}
 
 			var messageAsString = "[RunMigrations] start migration failed\n " +