+++ history.md
@@ -42,6 +42,14 @@ Semantic Versioning 2.0.0 is from version 0.1.6+
 ## List of versions
 
 DROP TABLE QueueItems;
+ALTER TABLE FileIndex
+DROP COLUMN ImageClassificationGeneratedAt;
+ALTER TABLE FileIndex
+DROP COLUMN ImageClassificationModel;
+ALTER TABLE FileIndex
+DROP COLUMN RejectedTags;
+ALTER TABLE FileIndex
+DROP COLUMN SuggestedTags;
 
 ## version 0.7.17 - 2026-05-04 {#v0.7.17}